Move QUACK epilogue analysis to private module
