+++ torch/_inductor/lowering.py
@@ -28,8 +28,8 @@ from torch._higher_order_ops.gemm_epilogue import (
     _gemm_epilogue_fusion,
     _SUPPORTED_GEMM_OP_NAMES,
     GEMM_EPILOGUE_OPS,
-    materialize_quack_epilogue,
 )
+from torch._higher_order_ops.gemm_epilogue_quack import materialize_quack_epilogue
 from torch._higher_order_ops.hints_wrap import hints_wrapper
 from torch._higher_order_ops.triton_kernel_wrap import triton_kernel_wrapper_mutation
 from torch._library.fake_class_registry import FakeScriptObject
